move ffi data struct
